requirements_dev.txt
homeassistant

black
flake8
pre-commit
pylint
python-dotenv
